Fix string encoding size check
The previously set limit of MaxInt16 was based on "Maximum n value is 32767" on wiki.vg/Protocol, but it appears this limit is not enforced in some cases.

For example, the JSON size of server query responses can grow past this limit, thus triggering the out-of-bounds error on large responses (e.g. modded forge servers sending mod lists for compat checks)

Fixes #15